+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/runtimes/TyeServiceIngressRuntime.kt
@@ -1,0 +1,6 @@
+package com.github.rafaelldi.tyeplugin.runtimes
+
+import com.github.rafaelldi.tyeplugin.model.TyeIngressService
+
+class TyeServiceIngressRuntime(service: TyeIngressService, parentRuntime: TyeApplicationRuntime) :
+    TyeServiceRuntime<TyeIngressService>(service, parentRuntime)